% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syllabus_dates.R
\name{class_date}
\alias{class_date}
\title{Return the date of a given class number}
\usage{
class_date(class_num, base = base_date, pattern, readable = TRUE)
}
\arguments{
\item{class_num}{The number of the class, starting at 1}

\item{base}{The starting date for the class}

\item{pattern}{See the function `make_next_class_date`}

\item{readable}{if TRUE, return a string; if FALSE, return a date object (defaults to TRUE)}
}
\value{
Either a string or a date object with the date of the `class_num`th class
}
\description{
Return the date of a given class number
}
