% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syllabus_dates.R
\name{make_next_class_date}
\alias{make_next_class_date}
\title{Create a function that returns date of subsequent classes each time it's called}
\usage{
make_next_class_date(base = base_date, pattern)
}
\arguments{
\item{base}{The date of the first class}

\item{pattern}{See `make_class_date`}
}
\value{
A function that, when called, returns the date of the next class
}
\description{
Create a function that returns date of subsequent classes each time it's called
}
