% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syllabus_dates.R
\name{week_num}
\alias{week_num}
\title{Given a date, calculate which week of class it takes place in}
\usage{
week_num(date, base = base_date)
}
\arguments{
\item{date}{The date}

\item{base}{The date of the first class}
}
\value{
The week in which the given class takes place. (The week with the first class is week 1)
}
\description{
Given a date, calculate which week of class it takes place in
}
